% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAge.R
\name{calcAge}
\alias{calcAge}
\title{Calculate the difference between two date in terms of unit of time.}
\usage{
calcAge(birthday, targetdate = Sys.Date(), unit = "year")
}
\arguments{
\item{birthday}{a character or character vector specifying birthday or base date.}

\item{targetdate}{a character or character vector specifying target date.}

\item{unit}{a character specifying unit for calculating the difference between the two dates. Values of "year", "month" and "day" are supported.}
}
\value{
a vector of age
}
\description{
This function calculate the difference between two date in terms of unit of time, and age can be obtained when \sQuote{year} is specified as unit argument.
}
\examples{
calcAge("1963-2-3")
}

