% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epifit-package.R
\docType{package}
\name{epifit-package}
\alias{epifit-package}
\alias{epifit-pkg}
\title{Flexible Modelling Functions for Epidemiological Data Analysis}
\description{
Provides flexible model fitting used in epidemiological data analysis
by a unified model specification, along with some data manipulation functions.
This package covers fitting of variety models including Cox regression models,
linear regression models, Poisson regression models, logistic models and
others whose likelihood is expressed in negative binomial, gamma and Weibull
distributions.
}
\author{
Author: Kazutaka DOI, Kei SAKABE and Masataka TAGURI
Maintainer: Kazutaka DOI \email{kztkdi@gmail.com}
}
\seealso{
\code{\link{epifit}}
}
\keyword{models}

