% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.R
\name{modes}
\alias{modes}
\title{Get the type or storage mode of an object.}
\usage{
modes(x)
}
\arguments{
\item{x}{any R object.}
}
\description{
Get the type or storage mode of an object.
}
\details{
This function has the almost same functionality as storage.mode except for recognition of factor. Intended for inner use in \code{showContents}, which must be able to be specified outside of \code{showContents} function.
}
\seealso{
\code{\link{showContents}}, \code{\link[base]{storage.mode}}.
}

