% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epifit.R
\name{print.epifit}
\alias{print.epifit}
\title{Print function for epifit object.}
\usage{
\method{print}{epifit}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{x}{Object of class \code{epifit}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to be printed in values. The default,  uses \code{max(\link[base:options]{getOption}}(digits - 4, 3)).}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function print result of function \code{\link{epifit}}
}

